rename user.entity
