+++ src/modules/users/users.provider.ts
@@ -1,6 +1,6 @@
 import { Provider } from '@nestjs/common';
 
-import { User } from './users.entity';
+import { User } from './user.entity';
 import { USER_REPOSITORY } from '../../core/constants';
 
 export const usersProvider: Provider[] = [
